+++ supervisely/nn/benchmark/__init__.py
@@ -1,1 +1,2 @@
 from supervisely.nn.benchmark.benchmark import Benchmark
+from supervisely.nn.benchmark.object_detection_benchmark import ObjectDetectionBenchmark
