+++ supervisely/nn/benchmark/evaluation/__init__.py
@@ -1,3 +1,2 @@
 from supervisely.nn.benchmark.evaluation.base_evaluator import BaseEvaluator
-from supervisely.nn.benchmark.evaluation.base_benchmark import BaseBenchmark
-from supervisely.nn.benchmark.evaluation.object_detection.evaluator import ObjectDetectionEvaluator
+from supervisely.nn.benchmark.evaluation.object_detection_evaluator import ObjectDetectionEvaluator
