+++ supervisely/nn/benchmark/evaluation/object_detection/__init__.py
@@ -1,4 +1,3 @@
 from supervisely.nn.benchmark.evaluation.object_detection.calculate_metrics import calculate_metrics
 from supervisely.nn.benchmark.evaluation.object_detection.metric_provider import MetricProvider
-from supervisely.nn.benchmark.evaluation.object_detection.evaluator import ObjectDetectionEvaluator
-from supervisely.nn.benchmark.evaluation.object_detection.benchmark import ObjectDetectionBenchmark
+from supervisely.nn.benchmark.evaluation.object_detection_evaluator import ObjectDetectionEvaluator
